updated EditTests & CrearOrden

+++ client/src/views/EditTests.vue
@@ -302,32 +302,59 @@
 			) {
 				alert("Por Favor Ingresar Datos Validos");
 			} else {
-				selectedPerfil.value.name = nombrePerfilNuevo.value.value;
-				selectedPerfil.value.cost_bs = costoBsPerfilNuevo.value.value;
-				selectedPerfil.value.cost_usd = costoDolaresPerfilNuevo.value.value;
+				const perfilesMenosSeleccionado = perfiles.value.filter((perfil) => {
+					return perfil.idProfile != selectedPerfil.value.idProfile;
+				});
 
-				tests.updateProfile(selectedPerfil.value.idProfile, selectedPerfil.value);
-				if (campos.value.length) {
-					await tests.createInputs(selectedPerfil.value.idProfile, campos.value);
-				}
+				if (
+					perfilesMenosSeleccionado.some((item) => {
+						const nombrePerfilExistente = item.name
+							.normalize("NFD")
+							.replace(/[\u0300-\u036f]/g, "")
+							.replace(/\s+/g, " ")
+							.trim()
+							.toLowerCase();
 
-				if (idCamposAgregados.value.length) {
-					console.log("campos agregados");
-					tests.createInputsInProfile(selectedPerfil.value.idProfile, idCamposAgregados.value);
-				}
+						const nombrePerfilEditadoLimpiado =
+							nombrePerfilNuevo.value.value
+								?.normalize("NFD")
+								.replace(/[\u0300-\u036f]/g, "")
+								.replace(/\s+/g, " ")
+								.trim()
+								.toLowerCase() || "";
 
-				if (idCamposAgregados.value.length) {
-					console.log("campos agregados");
-					tests.createInputsInProfile(selectedPerfil.value.idProfile, idCamposAgregados.value);
-				}
+						return nombrePerfilExistente === nombrePerfilEditadoLimpiado;
+					})
+				) {
+					showToast("Perfil Ya Existe", "warning", alertCircleOutline);
+				} else {
+					selectedPerfil.value.name = nombrePerfilNuevo.value.value;
+					selectedPerfil.value.cost_bs = costoBsPerfilNuevo.value.value;
+					selectedPerfil.value.cost_usd = costoDolaresPerfilNuevo.value.value;
 
-				if (idCamposEliminados.value.length) {
-					console.log("campos eliminados");
-					tests.deleteInputsInProfile(selectedPerfil.value.idProfile, idCamposEliminados.value);
+					tests.updateProfile(selectedPerfil.value.idProfile, selectedPerfil.value);
+					if (campos.value.length) {
+						await tests.createInputs(selectedPerfil.value.idProfile, campos.value);
+					}
+
+					if (idCamposAgregados.value.length) {
+						console.log("campos agregados");
+						tests.createInputsInProfile(selectedPerfil.value.idProfile, idCamposAgregados.value);
+					}
+
+					if (idCamposAgregados.value.length) {
+						console.log("campos agregados");
+						tests.createInputsInProfile(selectedPerfil.value.idProfile, idCamposAgregados.value);
+					}
+
+					if (idCamposEliminados.value.length) {
+						console.log("campos eliminados");
+						tests.deleteInputsInProfile(selectedPerfil.value.idProfile, idCamposEliminados.value);
+					}
+					showToast("Perfil Actualizado Exitosamente!", "creado", checkboxOutline);
+					update.value = false;
+					crearCampo.value = false;
 				}
-				showToast("Perfil Actualizado Exitosamente!", "creado", checkboxOutline);
-				update.value = false;
-				crearCampo.value = false;
 			}
 		}
 	};
